src/xdeid3d/gui/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Stable Diffusion (for text-to-3D generation)
diffusers>=0.35.0
transformers>=4.57.0
accelerate>=1.11.0

# Image processing
Pillow>=10.2.0
opencv-python>=4.9.0

# Deep learning
torch>=2.9.0
torchvision>=0.24.0

# Numerical computing
numpy>=1.24.0
scipy>=1.10.0

# 3D mesh processing
trimesh>=4.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
